extract store implementation to abstract store
